Fix a couple easy errors in animview
Errors reported by cppcheck
* undefined behaviour shifting a signed integer.
* memory leak if rnc length check fails.

+++ AnimView/th.h
@@ -117,7 +117,7 @@ public:
         if(0 <= iLayer && iLayer < 13)
             for(int iId = 0; iId < 32; ++iId)
             {
-                if((m_iMask[iLayer] & (1 << iId)) != 0)
+                if((m_iMask[iLayer] & (static_cast<std::uint32_t>(1) << iId)) != 0)
                     return true;
             }
         return false;
